requirements.txt
backcall==0.1.0
decorator==4.3.0
Django==2.1.1
django-bootstrap4==0.0.7
ipython==6.5.0
ipython-genutils==0.2.0
jedi==0.12.1
lxml==4.2.5
nltk==3.3
parso==0.3.1
pexpect==4.6.0
pickleshare==0.7.4
pkg-resources==0.0.0
prompt-toolkit==1.0.15
ptyprocess==0.6.0
Pygments==2.2.0
python-docx==0.8.7
pytz==2018.5
simplegeneric==0.8.1
six==1.11.0
traitlets==4.3.2
Unidecode==1.0.22
wcwidth==0.1.7
